Refactor CaptureUserOut() into util so it can be used elsewhere
