+++ pkg/ddevapp/providerDrudS3_test.go
@@ -54,7 +54,7 @@ func TestDrudS3ConfigCommand(t *testing.T) {
 	scanner := bufio.NewScanner(strings.NewReader(input))
 	util.SetInputScanner(scanner)
 
-	restoreOutput := testcommon.CaptureUserOut()
+	restoreOutput := util.CaptureUserOut()
 	err = app.PromptForConfig()
 	assert.NoError(err)
 	_ = restoreOutput()
@@ -75,7 +75,7 @@ func TestDrudS3ConfigCommand(t *testing.T) {
 	scanner = bufio.NewScanner(strings.NewReader(input))
 	util.SetInputScanner(scanner)
 
-	restoreOutput = testcommon.CaptureUserOut()
+	restoreOutput = util.CaptureUserOut()
 	err = app.PromptForConfig()
 	assert.NoError(err)
 	_ = restoreOutput()
@@ -95,7 +95,7 @@ func TestDrudS3ConfigCommand(t *testing.T) {
 	scanner = bufio.NewScanner(strings.NewReader(input))
 	util.SetInputScanner(scanner)
 
-	restoreOutput = testcommon.CaptureUserOut()
+	restoreOutput = util.CaptureUserOut()
 	err = app.PromptForConfig()
 	assert.Error(err)
 	if err != nil {
@@ -109,7 +109,7 @@ func TestDrudS3ConfigCommand(t *testing.T) {
 	scanner = bufio.NewScanner(strings.NewReader(input))
 	util.SetInputScanner(scanner)
 
-	restoreOutput = testcommon.CaptureUserOut()
+	restoreOutput = util.CaptureUserOut()
 	err = app.PromptForConfig()
 	assert.NoError(err)
